+++ inc/bkgd.h
@@ -36,7 +36,7 @@ struct LITE
 {
     BMAT34 bmat34;
     BRS rIntensity;
-    long lt; // light type
+    int32_t lt; // light type
 };
 VERIFY_STRUCT_SIZE(LITE, 56);
 const BOM kbomLite = 0xfffffff0;
@@ -89,7 +89,7 @@ struct BDS
 {
     int16_t bo;
     int16_t osk;
-    long vlm;
+    int32_t vlm;
     bool fLoop;
     TAG tagSnd;
 };
@@ -111,19 +111,19 @@ class BKGD : public BKGD_PAR
   protected:
     BACT *_prgbactLight; // array of light br_actors
     BLIT *_prgblitLight; // array of light data
-    long _cbactLight;    // count of lights
+    int32_t _cbactLight; // count of lights
     bool _fLites;        // lights are on
     bool _fLeaveLitesOn; // Don't turn out the lights
-    long _ccam;          // count of cameras in this background
-    long _icam;          // current camera
+    int32_t _ccam;       // count of cameras in this background
+    int32_t _icam;       // current camera
     BMAT34 _bmat34Mouse; // camera matrix for mouse model
     BRA _braRotY;        // Y rotation of current camera
     CNO _cnoSnd;         // background sound
     STN _stn;            // name of this background
     PGL _pglclr;         // palette for this background
     uint8_t _bIndexBase; // first index for palette
-    long _iaposLast;     // Last placement point we used
-    long _iaposNext;     // Next placement point to use
+    int32_t _iaposLast;  // Last placement point we used
+    int32_t _iaposNext;  // Next placement point to use
     PGL _pglapos;        // actor placement point(s) for current view
     BRS _xrPlace;
     BRS _yrPlace;
@@ -135,13 +135,13 @@ class BKGD : public BKGD_PAR
 
   protected:
     bool _FInit(PCFL pcfl, CTG ctg, CNO cno);
-    long _Ccam(PCFL pcfl, CTG ctg, CNO cno);
+    int32_t _Ccam(PCFL pcfl, CTG ctg, CNO cno);
     void _SetupLights(PGL pgllite);
 
   public:
     static bool FAddTagsToTagl(PTAG ptagBkgd, PTAGL ptagl);
     static bool FCacheToHD(PTAG ptagBkgd);
-    static bool FReadBkgd(PCRF pcrf, CTG ctg, CNO cno, PBLCK pblck, PBACO *ppbaco, long *pcb);
+    static bool FReadBkgd(PCRF pcrf, CTG ctg, CNO cno, PBLCK pblck, PBACO *ppbaco, int32_t *pcb);
     ~BKGD(void);
     void GetName(PSTN pstn);
 
@@ -156,15 +156,15 @@ class BKGD : public BKGD_PAR
         _fLeaveLitesOn = fLeaveLitesOn;
     }
 
-    long Ccam(void)
+    int32_t Ccam(void)
     {
         return _ccam;
     } // count of cameras in background
-    long Icam(void)
+    int32_t Icam(void)
     {
         return _icam;
     }                                        // currently selected camera
-    bool FSetCamera(PBWLD pbwld, long icam); // change camera to icam
+    bool FSetCamera(PBWLD pbwld, int32_t icam); // change camera to icam
 
     void GetMouseMatrix(BMAT34 *pbmat34);
     BRA BraRotYCamera(void)
@@ -174,14 +174,14 @@ class BKGD : public BKGD_PAR
     void GetActorPlacePoint(BRS *pxr, BRS *pyr, BRS *pzr);
     void ReuseActorPlacePoint(void);
 
-    void GetDefaultSound(PTAG ptagSnd, long *pvlm, bool *pfLoop)
+    void GetDefaultSound(PTAG ptagSnd, int32_t *pvlm, bool *pfLoop)
     {
         *ptagSnd = _bds.tagSnd;
         *pvlm = _bds.vlm;
         *pfLoop = _bds.fLoop;
     }
 
-    bool FGetPalette(PGL *ppglclr, long *piclrMin);
+    bool FGetPalette(PGL *ppglclr, int32_t *piclrMin);
     void GetCameraPos(BRS *pxr, BRS *pyr, BRS *pzr);
 
 #ifdef DEBUG
